Move http proxy configuration out of the package
Defining your proxy inside a package is a recipe for disaster.
Those things are best handled by configs.  In fact, we're actually NOT
setting a proxy in this package, we're adding some cruft that makes you
think we're doing it, when we're not.  Abdicate responsibility for setting
the proxy and remind the user to do it.