Remove naive npm download verification
As created, this RPM scripting did (pre-2020 / 2023) a `find | shasum` over the data downloaded from npm registries.
However, across versions, is it not uncommon for the package.json file of node_modules we download to change and be curated / tweaked with underscore-leading data from the registry, indicating highly variable items, including the path you used to download into the RPM packaging host.  Furthermore, the upstream ad-ldap-connector is not exact-version pinned, meaning each change would have to be hand-audited.
As such, package.json fluctuates and breaks between OS version, registry version, and sometimes even packaging runs on the same host.
This might've made sense in earlier supply-chain days, but it no longer does.  While the intention is smart and "make sure you can rebuild this consistently" the sheer weight of the number of packages here makes this utterly untenable, and the static checksum across the years means this approach is unworkable.

As such, the checksums are removed and "just build the thing" is the approach taken.